`PhotosSearch` now return all pages data.

+++ jsonstruct/jsonstruct.go
@@ -75,9 +75,9 @@ type Photos struct {
 // PhotosSearch in flickr.photos.search
 type PhotosSearch struct {
 	Photos struct {
-		Page    int64   `json:"page"`
-		Pages   int64   `json:"pages"`
-		Perpage int64   `json:"perpage"`
+		Page    int     `json:"page"`
+		Pages   int     `json:"pages"`
+		Perpage int     `json:"perpage"`
 		Total   string  `json:"total"`
 		Photo   []Photo `json:"photo"`
 	} `json:"photos"`
